+++ test/load-steps/playwright.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from '@playwright/test';
+import { defineBddConfig } from 'playwright-bdd';
+
+const testDir = defineBddConfig({
+  paths: ['features/*.feature'],
+  steps: 'steps/*',
+});
+
+export default defineConfig({
+  testDir,
+});