+++ test/load-steps/steps/index.ts
@@ -1,0 +1,5 @@
+import { createBdd } from 'playwright-bdd';
+
+const { When } = createBdd();
+
+When('I create todo', async () => {});