+++ test/unit/test.mjs
@@ -1,0 +1,23 @@
+import path from 'node:path';
+import { test, expect, TestDir } from '../_helpers/index.mjs';
+
+const testDir = new TestDir(import.meta);
+
+test(`${testDir.name} (finalizePattern)`, async () => {
+  const modulePath = path.resolve('dist/utils/paths.js');
+  const { finalizePattern } = await import(modulePath);
+
+  expect(finalizePattern('a/steps', 'js')).toEqual('a/steps/**/*.js');
+  expect(finalizePattern('a/steps/', 'js')).toEqual('a/steps/**/*.js');
+  expect(finalizePattern('a/steps//', 'js')).toEqual('a/steps/**/*.js');
+  expect(finalizePattern('a/steps/**', 'js')).toEqual('a/steps/**/*.js');
+  expect(finalizePattern('a/steps/*', 'js')).toEqual('a/steps/*.js');
+  expect(finalizePattern('a/steps/*.js', 'js')).toEqual('a/steps/*.js');
+  expect(finalizePattern('a/steps/**/*.js', 'js')).toEqual('a/steps/**/*.js');
+
+  // backslash (on win)
+  if (path.sep === '\\') {
+    expect(finalizePattern('a\\steps', 'js')).toEqual('a/steps/**/*.js');
+    expect(finalizePattern('a\\steps\\**', 'js')).toEqual('a/steps/**/*.js');
+  }
+});